fix: rename folder to internal
