+++ internal/core/ports/email_ports.go
