+++ internal/core/ports/user_port.go
@@ -1,7 +1,7 @@
 package ports
 
 import (
-	"github.com/PitiNarak/condormhub-backend/internals/core/domain"
+	"github.com/PitiNarak/condormhub-backend/internal/core/domain"
 	"github.com/gofiber/fiber/v2"
 	"github.com/google/uuid"
 )
