+++ internal/core/services/email_sender.go
@@ -3,7 +3,7 @@ package services
 import (
 	"fmt"
 
-	"github.com/PitiNarak/condormhub-backend/internals/core/ports"
+	"github.com/PitiNarak/condormhub-backend/internal/core/ports"
 	"github.com/PitiNarak/condormhub-backend/pkg/utils"
 	"github.com/go-gomail/gomail"
 	"github.com/google/uuid"
