+++ internal/core/services/user_service.go
@@ -3,8 +3,8 @@ package services
 import (
 	"errors"
 
-	"github.com/PitiNarak/condormhub-backend/internals/core/domain"
-	"github.com/PitiNarak/condormhub-backend/internals/core/ports"
+	"github.com/PitiNarak/condormhub-backend/internal/core/domain"
+	"github.com/PitiNarak/condormhub-backend/internal/core/ports"
 	"github.com/PitiNarak/condormhub-backend/pkg/utils"
 	"github.com/google/uuid"
 	"golang.org/x/crypto/bcrypt"
