+++ internal/handlers/sample_log_handler.go
@@ -1,8 +1,8 @@
 package handlers
 
 import (
-	"github.com/PitiNarak/condormhub-backend/internals/core/domain"
-	"github.com/PitiNarak/condormhub-backend/internals/core/ports"
+	"github.com/PitiNarak/condormhub-backend/internal/core/domain"
+	"github.com/PitiNarak/condormhub-backend/internal/core/ports"
 	"github.com/go-playground/validator"
 	"github.com/gofiber/fiber/v2"
 	"github.com/google/uuid"
